Chain predictions off versions
This connects predictions to versions conceptually, to match their
concept elsewhere, and gives us a clear API for differentiating from
other kinds of predictions in the future.

We create `Version` and `Prediction` classes we don't expect to expose
externally. The entrypoint for users continues to be methods called from
`ReplicateClient` directly.