Use "version" not "version" as the noun
We specify a model by the pattern `owner/model@version`.

Conceptually, the model is the thing, not the version. The version is a
way of specifying which version of the model we want, and doesn't make
much sense in isolation (even if the API itself doesn't ask for the
model (yet)).