Integration tests for common runtimes
This commit adds some basic integration tests and a new ci workflow
for our three common runtimes, nodejs with commonjs and esm as well
as TypeScript. The aim here is to ensure that the interface is
consistent even if we refactor/restructure the package.

The change does require a REPLICATE_API_TOKEN to be available in
the CI environment.