Add failing cloudflare integration suite
This is currently failing but acts as a starting point for verifying that
we can run the Replicate library on Cloudflare workers.

The test creates a cloudflare worker that uses the streaming API (as this
exercises a bunch of the API surface area) and uses the experimental
unstable_dev library to run the worker in a test. We then verify that
the test receives a successful response.

Currently we have the following issues:

* `node:crypto` needs to be replaced with web crypto (node >= 15).
* replace Buffer with Blob where possible or abstract behind a conditional.
* `process.env` also needs to be accessed behind a conditional.