Remove usage of webcrypto from `node:crypto` in Node 18
The original code was intended to shim in support for the webcrypto
interface in Node 18, which was included indirectly as `webcrypto`
in the "node:crypto" module or globally via the
`--no-experimental-global-webcrypto` flag.

This change has caused many issues with bundlers and static analyzers
which do not like the obfuscated call to `require()`.

Node 18 will no longer receive security support as of 30 April 2025[1]
and as such it feels like we can now drop this workaround in favor of
documenting alternative approaches.

[1]: https://endoflife.date/nodejs