Fix url key of createFileOutput options for streaming (#350)
Fix bug with broken `FileOutput` object being created as part of the streaming API as the wrong value for URL was passed to the constructor. This also extends the `replicate.stream()` interface to accept a `useFileOutput` configuration object.

Lastly, we now use the stream URL itself to detect if we should convert URL objects into `FileOutput`. 

---------

Co-authored-by: Aron Carroll <aron@replicate.com>